requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
pycparser==2.23
requests==2.32.5
urllib3==2.5.0
PyYAML==6.0.1
Flask==3.0.0
python-dotenv
PyJWT
